create DisciplineFSuite tests
